Revert temp storage element type back to std::size_t in launch_env
device_side_memory_resource::allocate() offsets are byte-addressed
via char* regardless of the vector's element type, so the size_t
element type wasn't a unit bug — it provided 8x headroom over the
declared expected_bytes_allocated. An exactly-sized uint8_t buffer
has no slack to absorb any small discrepancy between an algorithm's
actual temp-storage usage and the test's declared expectation, so
restore the size_t buffer.

+++ cub/test/catch2_test_launch_helper.h
@@ -339,7 +339,7 @@ void launch_env(ActionT action, Args... args)
   const size_t expected_bytes_allocated = env.query(get_expected_allocation_size_t{});
 
   c2h::device_vector<cudaError_t> d_error(1, cudaErrorInvalidValue);
-  c2h::device_vector<std::uint8_t> d_temp_storage(expected_bytes_allocated);
+  c2h::device_vector<std::size_t> d_temp_storage(expected_bytes_allocated);
   c2h::device_vector<std::size_t> d_allocated(1, 0);
   c2h::device_vector<std::size_t> d_deallocated(1, 0);
